[ios][ui] refreshable modifer moved into seperate file
